WorkerThread 파일로 분리
